cleanup unused files and reorganize
