+++ react/src/pages/components/GlobalFooter.tsx
