+++ react/src/pages/components/SectionTitle.tsx
